Relocate reserved identifier helper into plugin sources
Move loadReservedIdentifierNames from src/shared/identifier-metadata.js to src/plugin/src/reserved-identifiers.js so the plugin-only metadata loader lives beside the features that use it; update imports and tests for the new path.
